Move dtos into an own file and use them in the message_interface
